stubby-smash: some cleanups
There are a few cleanups here that I wanted when trying to use
stubby smash:

a. There is no real reason for 'bash', so change to be
   posix compliant and use /bin/sh shebang.  Using bash isn't
   wrong, but there is no need for or benefit from it here.
b. use tab consistently for indentation.  There was a mix of tab in
   some places (functions) and 8 spaces in the "main".
c. use 'command -v' rather than 'which'.  command -v is a shell
   builtin versus 'which' which is an external program.
d. give a better error message on bad input.  Instead of simply
   showing usage of non-existant file as input, show an error message
   identifying which argument was the problem.
e. use one argument "--long-option=value" as opposed to two
   ("--long-option" "value") when invoking objcopy.  The former is
   more obvious to the reader.

Signed-off-by: Scott Moser <smoser@brickies.net>
[PM: subject line tweaks]
Signed-off-by: Paul Moore <pmoore2@cisco.com>