Allow kernel cmdline from arguments that are in known whitelist.
If there is no built-in kernel command line, then check that the
provided command line contains only allowed tokens.

If booted in secure mode and an invalid token is found, then
the loader will exit failure.  If booted in insecure mode, then
the loader will print a warning and continue.

Token checks are against a static list of words (allowed).
Words in the allowed list beginning with a '^' will be considered
allowed if the input token begins with the word.

For example, '^console=t' will match:

   console=ttyS0
   console=tty0

But will not match:

    console
    console=/dev/ttyS0

The allowed word 'console=ttyS0' will *only* match the exact
string, and will not match console=ttyS0,115200.