Consistently set runtime cmdline and cmdline_len in stubby main.
When UEFI shell invokes stubby (kernel.efi) directly (without using
shim), then the LoadOptionsSize will include a terminating NULL.
The CHAR16* LoadOptions will include a NULL at the end (as it would have
to).

In all other ways that we expect to be invoked, the LoadOptionsSize does
*not* include a terminating NULL, and the LoadOptions is not even NULL
terminated.

These are the ways that we can be called now, and set the
rt_cmdline_len and rt_cmdline consistently as a NULL terminated string
that where rt_cmdline_len does not include the NULL:

 1. nvram shim.efi kernel.efi runtime-options ..
 2. nvram kernel.efi runtime-options ..
 3. uefi-shell shim.efi kernel.efi runtime-options
 4. uefi-shell kernel.efi runtime-options

(Previously path '4' was broken).