Pull in shim's load-options.c for use of parse_load_options.
This pulls in load-options.c from the shim upstream repository.
This is pulled from shim commit
  a78673b19e98b4f000a2d8043e4bfdaab6d83a05.

The only changes here are the copying of functions
 StrnCaseCmp and is_all_nulls.

The reason is for use of parse_load_options which is has logic we
need to fix a bug where the LoadOptions differ based on how
the program was loaded.