Adjust parse_load_options behavior for our use.
parse_load_options would always modify the LoadOptions to include a
null separating the first word and remaining words.  The behavior
we wanted was to only have it remove the efi filename but otherwise
leave the LoadOptions intact.

This way we can pass the rest of the string off as the linux
command line directly.

Also included with that change are:
 * replacement of perror with Print
 * removal of dprint ('d' here was debug for shim)
 * removal of hexdumpat
 * removal of no longer needed globals.