Fix problem where efi filename was used in kernel cmdline
kernel.efi would appear on the kernel's /proc/cmdline
When a stubby-based kernel was executed from the EFI shell as:

    kernel.efi root=my-rootfs my-arguments

Things to be aware of:

a. this problem did not occur when invoked by shim even from EFI shell.

b. is not desireable as the kernel's name is not strictly intended
as kernel command line input.  Note that many systems such as
Ubuntu do currently boot efi with a /proc/cmdline like
'BOOT_IMAGE=/vmlinuz-5.15.0-27-generic root=UUID=....'

The real reason for this change is that this could be used as an attack
to get a command line parameter read by the kernel.  An attacker could
rename a signed 'kernel.efi' to "rootkit=yes-please" or
"debug-shell=true" and have the kernel read that parameter that otherwise
would have been ignored.

So, we had to either make the efi filename also restricted as
part of the allowed_list or not pass it through.
The change here is to not pass it through.