Remove efi name from the runtime command line if it is present.
The UEFI shell sets LoadOptions in a different way than either
shim or nvram.  If you execute "stubby.efi arg1 arg2" from EFI
shell, then the LoadOptions value line will include "stubby.efi".

The fix here is just to drop the first token of the command line
value if it ends in .efi or .EFI.  Not perfect, but functional
for the current need.

Fixes #19.