Publish separate artifact for stubby, results with tar.
It appears that all artifacts uploaded get zipped, which lose permission
info and potentially case.  The documented solution is to make a tarball
and let the uploader zip up the tarball.  So that is what is done
here with the test results.

Also changed is to add sbat.csv to the stubby.efi.  So you have the
pieces that would be required to smoosh a kernel.

Note, there is duplication of sbat.csv and stubby.efi in test-results
but they make up like 30k total.