Change github action for timeout of 120s and 4 threads.
Github runners have 2 cores, but thanks to the logic error,
we know that historically they have done really well with 8 threads
running on those 2 cores with no KVM.

Since we did get *some* timeouts previously we will set the number
of threads to 4 to hopefully strike a balance between reliability
and speed.  Leave the timeout to 120s but be explicit in the
github action.