Add test case and fix for sb, builtin, noruntime, nomarker case
First, document cmdline cases and fix built-in,noruntime,no-marker,sb

There are a lot of different scenarios on how we want to handle
cmdline values provided to shim.  The following table summarizes the
ones that matter and how we should handle them.

.-------------------------------------------------------------------.
| case | builtin | runtime | builtin_has_marker | sb  mode | status |
|------|---------|---------|--------------------|----------|--------|
|  a   |  True   | False   | False              | insecure | ok     |
|  b   |  True   | False   | True               | insecure | ok     |
|  c   |  True   | True    | False              | insecure | ok     |
|  d   |  True   | True    | True               | insecure | ok     |
|--------------- ---------------------------------------------------|
|  e   |  True   | False   | False              |   secure | ok     |
|  f   |  True   | False   | True               |   secure | ok     |
|  g   |  True   | True    | False              |   secure | fail   |
|  h   |  True   | True    | True               |   secure | ok     |
`-------------------------------------------------------------------'

The primary failure case we must prevent is case g, where we have
a built-in cmdline, a runtime cmdline has been provided, but the
marker is not provided; without such we cannot create the correct
cmdline that inserts the runtime values in the correct spot.

The code already handled that case, but it also failed in
case e where we have a builtin, no runtime and no marker under SB.
For case e, we now copy the builtin cmdline into the buffer so we
don't boot into a kernel with no cmdline provided.

Issues fixed:

- kcmdline.c:
  - check_cmdline was only called on runtime buffer, which may be empty
  - only add a space to part1 if we have a runtime buffer, otherwise we
    only work with the original builtin cmdline value
  - only check_cmdline on runtime if we have one
  - don't set status to EFI_SUCCESS by default in check_cmdline
  - don't allow empty cmdline strings, or all spaces (no tokens)
  - mark empty cmdline tests EFI_SECURITY_VIOLATION

test-get-cmdline.c
  - mark empty cmdline test EFI_SECURITY_VIOLATION
  - Add the test number to the failure message

- Add testcase for shim builtin cmdline with no runtime and no marker

Test out the path where kernel.efi has a built-in value, without a
marker, and no runtime args are supplied.

test/harness:
- fix booted_expected_cli value, there was a copy-n-paste bug when
  implementing booted_expected_cli.  The __docstring__ described what
  it should do but the code referenced 'runtime' in the error message
  and tdata dictionary.
- fix efi-shim mode by copying kernel.efi as grubx64.efi into esp

Signed-off-by: Ryan Harper <rharper@woxford.com>