Generate git version and include that in source output
The current GIT_VERSION value only works if you build from a git
repo.  Once we release on github, the tarballs that one can download
are not git repos and the git command to describe the version fail.

Generate the git version value and write to a file.  In source tarball
builds, the value will come from the file rather than git.

- Add 'show-version' as makefile target
- Add 'show-version' to github workflow
- Adjust make-tarball tool to insert the git-version file

Signed-off-by: Ryan Harper <rharper@woxford.com>