Find the OVMF needle in a haystack
The older focal ovmf build is not installable via apt any longer.
Further jammy and newer releases all have ovmf which has uefi-shell
disabled in secureboot which makes those releases not useful for
stubby testing.

I have found a needle in the haystack of ovmf builds somewhere in
the jammy release where we have 2M and 4M builds and secureboot is
not disabled.

I'm not sure if this build has the snakeoil key bug or not but
I'm hoping this is sufficient to get some testing that stubby itself
compiles and boots in some cases to confirm that changing the objcopy
flag creates a functioning stubby.

Signed-off-by: Ryan Harper <rharper@woxford.com>